Gamut and MidiChordDef changes etc.
Changed signature of MidiChordDef ctor from
MidiChordDef(int nPitches, int rootPitch,
    List<int> absolutePitchHierarchy, int velocity,
    int msDuration, bool hasChordOff)
to
MidiChordDef(int msDuration, Gamut gamut,
    int rootNotatedPitch, int nPitchesPerChord,
    Envelope ornamentEnvelope = null)
Replaced
MoritzStatics.GetAscendingPitches(int nPitches,
    int rootPitch, List<int> absolutePitchHierarchy)
by
Gamut.GetChord(int rootPitch, int nPitches)
Replaced constructor
BasicMidiChordDef(int nPitches, int rootPitch,
    List<int> absolutePitchHierarchy, int velocity,
    int msDuration, bool hasChordOff)
by
BasicMidiChordDef(int msDuration, Gamut gamut,
    int rootPitch, int density)
GetVelocityPerAbsolutePitch(...) moved from
MoritzStatics to Gamut.
Deleted MoritzStatics.VelocityFactors.
Updated TestBlocks and Tombeau1 main comment.

+++ Moritz.Spec/Trk.cs
@@ -243,7 +243,7 @@ namespace Moritz.Spec
         /// This function applies equally to all the BasicMidiChordDefs in this MidiChordDef. 
         /// </summary>
         /// <param name="velocityPerAbsolutePitch">A list of 12 velocity values (range [0..127] in order of absolute pitch</param>
-        public void SetVelocityPerAbsolutePitch(List<int> velocityPerAbsolutePitch)
+        public void SetVelocityPerAbsolutePitch(List<byte> velocityPerAbsolutePitch)
         {
             #region conditions
             Debug.Assert(velocityPerAbsolutePitch.Count == 12);
